Add files via upload
finished the linux section, was just a lot of copying and pasting with slight edits here and there from my windows part. whats left now is to just test both on VM's to see if all of the options work, i also have to brainstorm a way to safely store the key somewhere before encrypting all the user's files.